+++ app/features/ledger/ConcordiumLedgerClientMain.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type HwTransport from '@ledgerhq/hw-transport';
 import { Buffer } from 'buffer/';
-import { BrowserWindow } from 'electron';
+import EventEmitter from 'events';
 import { Transport, TransportImpl } from './Transport';
 import {
     getPublicKey,
@@ -44,28 +44,6 @@ import signAuthorizationKeysUpdate from './SignAuthorizationKeysUpdate';
 import signAddIdentityProviderTransaction from './SignAddIdentityProvider';
 import EmulatorTransport from './EmulatorTransport';
 
-export interface LedgerIpcMessage<T> {
-    result?: T;
-    error?: any;
-}
-
-async function wrapResult<T>(
-    call: (...input: any[]) => Promise<T>,
-    ...args: any[]
-): Promise<LedgerIpcMessage<T>> {
-    try {
-        const result: LedgerIpcMessage<T> = {
-            result: await call(...args),
-        };
-        return result;
-    } catch (e) {
-        const error: LedgerIpcMessage<T> = {
-            error: JSON.stringify(e),
-        };
-        return error;
-    }
-}
-
 /**
  * Concordium Ledger API.
  *
@@ -78,9 +56,9 @@ async function wrapResult<T>(
 export default class ConcordiumLedgerClientMain {
     transport: Transport;
 
-    mainWindow: BrowserWindow;
+    eventEmitter: EventEmitter;
 
-    constructor(mainWindow: BrowserWindow, transport?: HwTransport) {
+    constructor(eventEmitter: EventEmitter, transport?: HwTransport) {
         if (transport) {
             this.transport = new TransportImpl(transport);
         } else {
@@ -88,62 +66,58 @@ export default class ConcordiumLedgerClientMain {
             // Only to be used for testing, as the emulator is not secure in any way.
             this.transport = new EmulatorTransport();
         }
-        this.mainWindow = mainWindow;
+        this.eventEmitter = eventEmitter;
     }
 
     closeTransport(): Promise<void> {
         return this.transport.close();
     }
 
-    getPublicKey(path: number[]): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(getPublicKey, this.transport, path);
+    getPublicKey(path: number[]): Promise<Buffer> {
+        return getPublicKey(this.transport, path);
     }
 
-    getPublicKeySilent(path: number[]): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(getPublicKeySilent, this.transport, path);
+    getPublicKeySilent(path: number[]): Promise<Buffer> {
+        return getPublicKeySilent(this.transport, path);
     }
 
-    getSignedPublicKey(
-        path: number[]
-    ): Promise<LedgerIpcMessage<SignedPublicKey>> {
-        return wrapResult(getSignedPublicKey, this.transport, path);
+    getSignedPublicKey(path: number[]): Promise<SignedPublicKey> {
+        return getSignedPublicKey(this.transport, path);
     }
 
-    getIdCredSec(identity: number): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(getIdCredSec, this.transport, identity);
+    getIdCredSec(identity: number): Promise<Buffer> {
+        return getIdCredSec(this.transport, identity);
     }
 
-    getPrfKey(identity: number): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(getPrfKey, this.transport, identity);
+    getPrfKey(identity: number): Promise<Buffer> {
+        return getPrfKey(this.transport, identity);
     }
 
     signTransfer(
         transaction: AccountTransaction,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(signTransfer, this.transport, path, transaction);
+    ): Promise<Buffer> {
+        return signTransfer(this.transport, path, transaction);
     }
 
     signUpdateCredentialTransaction(
         transaction: UpdateAccountCredentials,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signUpdateCredentialTransaction,
+    ): Promise<Buffer> {
+        return signUpdateCredentialTransaction(
             this.transport,
             path,
             transaction,
-            this.mainWindow
+            this.eventEmitter
         );
     }
 
     signPublicInformationForIp(
         publicInfoForIp: PublicInformationForIp,
         accountPathInput: AccountPathInput
-    ): Promise<LedgerIpcMessage<Buffer>> {
+    ): Promise<Buffer> {
         const accountPath = getAccountPath(accountPathInput);
-        return wrapResult(
-            signPublicInformationForIp,
+        return signPublicInformationForIp(
             this.transport,
             accountPath,
             publicInfoForIp
@@ -154,9 +128,8 @@ export default class ConcordiumLedgerClientMain {
         credentialDeployment: UnsignedCredentialDeploymentInformation,
         address: string,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signCredentialDeploymentOnExistingAccount,
+    ): Promise<Buffer> {
+        return signCredentialDeploymentOnExistingAccount(
             this.transport,
             credentialDeployment,
             address,
@@ -168,9 +141,8 @@ export default class ConcordiumLedgerClientMain {
         credentialDeployment: UnsignedCredentialDeploymentInformation,
         expiry: bigint,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signCredentialDeploymentOnNewAccount,
+    ): Promise<Buffer> {
+        return signCredentialDeploymentOnNewAccount(
             this.transport,
             credentialDeployment,
             expiry,
@@ -182,9 +154,8 @@ export default class ConcordiumLedgerClientMain {
         transaction: UpdateInstruction<ExchangeRate>,
         serializedPayload: Buffer,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signUpdateTransaction,
+    ): Promise<Buffer> {
+        return signUpdateTransaction(
             this.transport,
             0x06,
             path,
@@ -197,9 +168,8 @@ export default class ConcordiumLedgerClientMain {
         transaction: UpdateInstruction<ExchangeRate>,
         serializedPayload: Buffer,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signUpdateTransaction,
+    ): Promise<Buffer> {
+        return signUpdateTransaction(
             this.transport,
             0x06,
             path,
@@ -212,9 +182,8 @@ export default class ConcordiumLedgerClientMain {
         transaction: UpdateInstruction<TransactionFeeDistribution>,
         serializedPayload: Buffer,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signUpdateTransaction,
+    ): Promise<Buffer> {
+        return signUpdateTransaction(
             this.transport,
             0x22,
             path,
@@ -227,9 +196,8 @@ export default class ConcordiumLedgerClientMain {
         transaction: UpdateInstruction<FoundationAccount>,
         serializedPayload: Buffer,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signUpdateTransaction,
+    ): Promise<Buffer> {
+        return signUpdateTransaction(
             this.transport,
             0x24,
             path,
@@ -242,9 +210,8 @@ export default class ConcordiumLedgerClientMain {
         transaction: UpdateInstruction<MintDistribution>,
         serializedPayload: Buffer,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signUpdateTransaction,
+    ): Promise<Buffer> {
+        return signUpdateTransaction(
             this.transport,
             0x25,
             path,
@@ -257,9 +224,8 @@ export default class ConcordiumLedgerClientMain {
         transaction: UpdateInstruction<ProtocolUpdate>,
         serializedPayload: Buffer,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signUpdateProtocolTransaction,
+    ): Promise<Buffer> {
+        return signUpdateProtocolTransaction(
             this.transport,
             path,
             transaction,
@@ -271,9 +237,8 @@ export default class ConcordiumLedgerClientMain {
         transaction: UpdateInstruction<GasRewards>,
         serializedPayload: Buffer,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signUpdateTransaction,
+    ): Promise<Buffer> {
+        return signUpdateTransaction(
             this.transport,
             0x23,
             path,
@@ -286,9 +251,8 @@ export default class ConcordiumLedgerClientMain {
         transaction: UpdateInstruction<BakerStakeThreshold>,
         serializedPayload: Buffer,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signUpdateTransaction,
+    ): Promise<Buffer> {
+        return signUpdateTransaction(
             this.transport,
             0x27,
             path,
@@ -301,9 +265,8 @@ export default class ConcordiumLedgerClientMain {
         transaction: UpdateInstruction<ElectionDifficulty>,
         serializedPayload: Buffer,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signUpdateTransaction,
+    ): Promise<Buffer> {
+        return signUpdateTransaction(
             this.transport,
             0x26,
             path,
@@ -317,9 +280,8 @@ export default class ConcordiumLedgerClientMain {
         serializedPayload: Buffer,
         path: number[],
         INS: number
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signHigherLevelKeyUpdate,
+    ): Promise<Buffer> {
+        return signHigherLevelKeyUpdate(
             this.transport,
             path,
             transaction,
@@ -333,9 +295,8 @@ export default class ConcordiumLedgerClientMain {
         serializedPayload: Buffer,
         path: number[],
         INS: number
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signAuthorizationKeysUpdate,
+    ): Promise<Buffer> {
+        return signAuthorizationKeysUpdate(
             this.transport,
             path,
             transaction,
@@ -348,9 +309,8 @@ export default class ConcordiumLedgerClientMain {
         transaction: UpdateInstruction<AddIdentityProvider>,
         serializedPayload: Buffer,
         path: number[]
-    ): Promise<LedgerIpcMessage<Buffer>> {
-        return wrapResult(
-            signAddIdentityProviderTransaction,
+    ): Promise<Buffer> {
+        return signAddIdentityProviderTransaction(
             this.transport,
             path,
             transaction,
@@ -358,7 +318,7 @@ export default class ConcordiumLedgerClientMain {
         );
     }
 
-    getAppAndVersion(): Promise<LedgerIpcMessage<AppAndVersion>> {
-        return wrapResult(getAppAndVersion, this.transport);
+    getAppAndVersion(): Promise<AppAndVersion> {
+        return getAppAndVersion(this.transport);
     }
 }