+++ app/preload/database/identityDao.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { IpcMain } from 'electron';
+import type { Knex } from 'knex';
 import { knex } from '~/database/knex';
 import {
     identitiesTable,
@@ -7,15 +7,23 @@ import {
     credentialsTable,
     addressBookTable,
 } from '~/constants/databaseNames.json';
-import ipcCommands from '~/constants/ipcCommands.json';
+import { removeInitialAccount } from './accountDao';
 import {
     Account,
+    Credential,
     AccountStatus,
     AddressBookEntry,
     Identity,
     IdentityStatus,
 } from '~/utils/types';
+import { IdentityMethods } from '~/preload/preloadTypes';
 
+/**
+ * Get the identity number to be used to create the next identity with
+ * the wallet with the given id.
+ * @param walletId the database id key for the wallet used
+ * @returns the id for the next identity to be created by the given wallet
+ */
 export async function getNextIdentityNumber(walletId: number): Promise<number> {
     const model = (await knex())
         .table(identitiesTable)
@@ -35,8 +43,17 @@ async function updateIdentity(
     return (await knex())(identitiesTable).where({ id }).update(updatedValues);
 }
 
+/**
+ * Find all the identities for a given wallet.
+ * @returns a list of identities that have been created from the supplied wallet
+ */
 async function getIdentitiesForWallet(walletId: number): Promise<Identity[]> {
     return (await knex()).select().table(identitiesTable).where({ walletId });
+}
+
+async function removeIdentity(id: number, trx: Knex.Transaction) {
+    const table = (await knex())(identitiesTable).transacting(trx);
+    return table.where({ id }).del();
 }
 
 /**
@@ -44,12 +61,14 @@ async function getIdentitiesForWallet(walletId: number): Promise<Identity[]> {
  * initial account.
  * @param identityId the identity to reject
  */
-async function rejectIdentityAndDeleteInitialAccount(identityId: number) {
+async function rejectIdentityAndInitialAccount(identityId: number) {
     (await knex()).transaction(async (trx) => {
         await trx(identitiesTable)
             .where({ id: identityId })
             .update({ status: IdentityStatus.Rejected });
-        await trx(accountsTable).where({ identityId, isInitial: 1 }).del();
+        await trx(accountsTable)
+            .where({ identityId, isInitial: 1 })
+            .update({ status: AccountStatus.Rejected });
     });
 }
 
@@ -57,7 +76,7 @@ async function rejectIdentityAndDeleteInitialAccount(identityId: number) {
  * Transactionally inserts an identity and its initial account.
  * @returns the identityId of the inserted identity
  */
-async function addPendingIdentityAndInitialAccount(
+async function insertPendingIdentityAndInitialAccount(
     identity: Partial<Identity>,
     initialAccount: Omit<Account, 'identityId'>
 ) {
@@ -107,73 +126,27 @@ async function confirmIdentity(
     });
 }
 
-export default function initializeIpcHandlers(ipcMain: IpcMain) {
-    ipcMain.handle(
-        ipcCommands.database.identity.insert,
-        async (_event, identity: Identity) => {
-            return insertIdentity(identity);
-        }
-    );
-
-    ipcMain.handle(
-        ipcCommands.database.identity.update,
-        async (_event, id: number, updatedValues: Record<string, unknown>) => {
-            return updateIdentity(id, updatedValues);
-        }
-    );
-
-    ipcMain.handle(
-        ipcCommands.database.identity.getIdentitiesForWallet,
-        async (_event, walletId: number) => {
-            return getIdentitiesForWallet(walletId);
-        }
-    );
-
-    ipcMain.handle(
-        ipcCommands.database.identity.getNextIdentityNumber,
-        async (_event, walletId: number) => {
-            return getNextIdentityNumber(walletId);
-        }
-    );
-
-    ipcMain.handle(
-        ipcCommands.database.identity.rejectIdentityAndDeleteInitialAccount,
-        async (_event, identityId: number) => {
-            return rejectIdentityAndDeleteInitialAccount(identityId);
-        }
-    );
-
-    ipcMain.handle(
-        ipcCommands.database.identity.confirmIdentity,
-        async (
-            _event,
-            identityId: number,
-            identityObjectJson: string,
-            accountAddress: string,
-            credential: Credential,
-            addressBookEntry: AddressBookEntry
-        ) => {
-            return confirmIdentity(
-                identityId,
-                identityObjectJson,
-                accountAddress,
-                credential,
-                addressBookEntry
-            );
-        }
-    );
-
-    ipcMain.handle(
-        ipcCommands.database.identity.insertPendingIdentityAndInitialAccount,
-        async (
-            _event,
-            identity: Partial<Identity>,
-            initialAccount: Omit<Account, 'identityId'>
-        ) => {
-            return addPendingIdentityAndInitialAccount(
-                identity,
-                initialAccount
-            );
-        }
-    );
+async function removeIdentityAndInitialAccount(identityId: number) {
+    const db = await knex();
+    await db.transaction((trx) => {
+        removeInitialAccount(identityId, trx)
+            .then(() => {
+                return removeIdentity(identityId, trx);
+            })
+            .then(trx.commit)
+            .catch(trx.rollback);
+    });
 }
+
+const exposedMethods: IdentityMethods = {
+    getNextIdentityNumber,
+    insert: insertIdentity,
+    update: updateIdentity,
+    getIdentitiesForWallet,
+    rejectIdentityAndInitialAccount,
+    removeIdentityAndInitialAccount,
+    confirmIdentity,
+    insertPendingIdentityAndInitialAccount,
+};
+
+export default exposedMethods;