+++ app/preload/ledger/ledger.ts
@@ -1,6 +1,5 @@
-import { BrowserWindow, IpcMain } from 'electron';
 import { Buffer } from 'buffer/';
-import ledgerIpcCommands from '~/constants/ledgerIpcCommands.json';
+import EventEmitter from 'events';
 import { parse } from '~/utils/JSONHelper';
 import {
     closeTransport,
@@ -26,55 +25,26 @@ import {
     UpdateAccountCredentials,
     AddIdentityProvider,
 } from '~/utils/types';
+import { LedgerCommands } from '~/preload/preloadTypes';
 
-export default function initializeIpcHandlers(
-    ipcMain: IpcMain,
-    mainWindow: BrowserWindow
-) {
-    ipcMain.handle(
-        ledgerIpcCommands.getPublicKey,
-        (_event, keypath: number[]) => {
-            return getLedgerClient().getPublicKey(keypath);
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.getPublicKeySilent,
-        (_event, keypath: number[]) => {
-            return getLedgerClient().getPublicKeySilent(keypath);
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.getSignedPublicKey,
-        (_event, keypath: number[]) => {
-            return getLedgerClient().getSignedPublicKey(keypath);
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.getIdCredSec,
-        (_event, identity: number) => {
-            return getLedgerClient().getIdCredSec(identity);
-        }
-    );
-
-    ipcMain.handle(ledgerIpcCommands.getPrfKey, (_event, identity: number) => {
-        return getLedgerClient().getPrfKey(identity);
-    });
-
-    ipcMain.handle(
-        ledgerIpcCommands.signTransfer,
-        (_event, transactionAsJson: string, keypath: number[]) => {
+export default function exposedMethods(
+    eventEmitter: EventEmitter
+): LedgerCommands {
+    return {
+        getPublicKey: (keypath: number[]) =>
+            getLedgerClient().getPublicKey(keypath),
+        getPublicKeySilent: (keypath: number[]) =>
+            getLedgerClient().getPublicKeySilent(keypath),
+        getSignedPublicKey: (keypath: number[]) =>
+            getLedgerClient().getSignedPublicKey(keypath),
+        getIdCredSec: (identity: number) =>
+            getLedgerClient().getIdCredSec(identity),
+        getPrfKey: (identity: number) => getLedgerClient().getPrfKey(identity),
+        signTransfer: (transactionAsJson: string, keypath: number[]) => {
             const transaction: AccountTransaction = parse(transactionAsJson);
             return getLedgerClient().signTransfer(transaction, keypath);
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signPublicInformationForIp,
-        (
-            _event,
+        },
+        signPublicInformationForIp: (
             publicInfoForIp: PublicInformationForIp,
             accountPathInput: AccountPathInput
         ) => {
@@ -82,12 +52,11 @@ export default function initializeIpcHandlers(
                 publicInfoForIp,
                 accountPathInput
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signUpdateCredentialTransaction,
-        (_event, transactionAsJson: string, path: number[]) => {
+        },
+        signUpdateCredentialTransaction: (
+            transactionAsJson: string,
+            path: number[]
+        ) => {
             const transaction: UpdateAccountCredentials = parse(
                 transactionAsJson
             );
@@ -95,13 +64,8 @@ export default function initializeIpcHandlers(
                 transaction,
                 path
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signCredentialDeploymentOnExistingAccount,
-        (
-            _event,
+        },
+        signCredentialDeploymentOnExistingAccount: (
             credentialDeployment: UnsignedCredentialDeploymentInformation,
             address: string,
             keypath: number[]
@@ -111,13 +75,8 @@ export default function initializeIpcHandlers(
                 address,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signCredentialDeploymentOnNewAccount,
-        (
-            _event,
+        },
+        signCredentialDeploymentOnNewAccount: (
             credentialDeployment: UnsignedCredentialDeploymentInformation,
             expiry: string,
             keypath: number[]
@@ -127,13 +86,8 @@ export default function initializeIpcHandlers(
                 parse(expiry),
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signMicroGtuPerEuro,
-        (
-            _event,
+        },
+        signMicroGtuPerEuro: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[]
@@ -146,13 +100,8 @@ export default function initializeIpcHandlers(
                 serializedPayload,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signEuroPerEnergy,
-        (
-            _event,
+        },
+        signEuroPerEnergy: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[]
@@ -165,13 +114,8 @@ export default function initializeIpcHandlers(
                 serializedPayload,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signTransactionFeeDistribution,
-        (
-            _event,
+        },
+        signTransactionFeeDistribution: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[]
@@ -184,13 +128,8 @@ export default function initializeIpcHandlers(
                 serializedPayload,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signFoundationAccount,
-        (
-            _event,
+        },
+        signFoundationAccount: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[]
@@ -203,13 +142,8 @@ export default function initializeIpcHandlers(
                 serializedPayload,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signMintDistribution,
-        (
-            _event,
+        },
+        signMintDistribution: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[]
@@ -222,13 +156,8 @@ export default function initializeIpcHandlers(
                 serializedPayload,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signProtocolUpdate,
-        (
-            _event,
+        },
+        signProtocolUpdate: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[]
@@ -241,13 +170,8 @@ export default function initializeIpcHandlers(
                 serializedPayload,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signAddIdentityProvider,
-        (
-            _event,
+        },
+        signAddIdentityProvider: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[]
@@ -260,13 +184,8 @@ export default function initializeIpcHandlers(
                 serializedPayload,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signGasRewards,
-        (
-            _event,
+        },
+        signGasRewards: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[]
@@ -279,13 +198,8 @@ export default function initializeIpcHandlers(
                 serializedPayload,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signBakerStakeThreshold,
-        (
-            _event,
+        },
+        signBakerStakeThreshold: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[]
@@ -298,13 +212,8 @@ export default function initializeIpcHandlers(
                 serializedPayload,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signElectionDifficulty,
-        (
-            _event,
+        },
+        signElectionDifficulty: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[]
@@ -317,13 +226,8 @@ export default function initializeIpcHandlers(
                 serializedPayload,
                 keypath
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signHigherLevelKeysUpdate,
-        (
-            _event,
+        },
+        signHigherLevelKeysUpdate: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[],
@@ -338,13 +242,8 @@ export default function initializeIpcHandlers(
                 keypath,
                 INS
             );
-        }
-    );
-
-    ipcMain.handle(
-        ledgerIpcCommands.signAuthorizationKeysUpdate,
-        (
-            _event,
+        },
+        signAuthorizationKeysUpdate: (
             transactionAsJson: string,
             serializedPayload: Buffer,
             keypath: number[],
@@ -359,21 +258,9 @@ export default function initializeIpcHandlers(
                 keypath,
                 INS
             );
-        }
-    );
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    ipcMain.handle(ledgerIpcCommands.getAppAndVersion, (_event) => {
-        return getLedgerClient().getAppAndVersion();
-    });
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    ipcMain.handle(ledgerIpcCommands.subscribe, (_event) => {
-        return subscribeLedger(mainWindow);
-    });
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    ipcMain.handle(ledgerIpcCommands.closeTransport, (_event) => {
-        return closeTransport();
-    });
+        },
+        getAppAndVersion: () => getLedgerClient().getAppAndVersion(),
+        subscribe: () => subscribeLedger(eventEmitter),
+        closeTransport,
+    };
 }