+++ app/utils/WalletPairing.ts
@@ -1,4 +1,4 @@
-import { getId, insertWallet } from '~/database/WalletDao';
+import { getWalletId, insertWallet } from '~/database/WalletDao';
 import ConcordiumLedgerClient from '~/features/ledger/ConcordiumLedgerClient';
 import { getPairingPath } from '~/features/ledger/Path';
 import { setCurrentWalletId } from '~/features/WalletSlice';
@@ -17,7 +17,7 @@ export default async function pairWallet(
     const pairingKey = (
         await ledger.getPublicKeySilent(getPairingPath())
     ).toString('hex');
-    let walletId = await getId(pairingKey);
+    let walletId = await getWalletId(pairingKey);
     if (walletId === undefined) {
         walletId = await insertWallet(pairingKey, WalletType.LedgerNanoS);
         dispatch(setCurrentWalletId(walletId));