+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 -   It is now possible to rename accounts and identities.
 -   Added an option to add an address book entry while creating a transfer transaction.
 -   Added an introductory screen to set up a node connection for first time users.
+-   It is now possible to remove a failed identity.
 
 ### Changed